Fixed visual bug when loading test flowchart
Loading partway through the flowchart requires for character portraits to be displayed again. However, since portraits are shown on the first node, we don't run the Save Point Loaded node when loading the first node so this effect doesn't trigger. Testing should be done when designing a flow chart to make sure that visual effects don't play twice when transitioning between nodes or when loading a save.