Figuring out how to animate character sprites
Doable using the iTween Fungus commands, but we need to modify a way so that we can set before runtime which sprites to animate. The character sprite holder which is created at runtime needs to be affected by the iTween, but we don't have access to that before the game starts. To get around this,  possibly allow setting the target to be a Character GameObject, then, find the holder at runtime when running the iTween command.