Changed how the prep phase menu is shown to enable setting the scrim results using a custom Fungus command
Prep phase menu is now displayed by setting alpha and interactivity values on a CanvasGroup component instead of enabling/disabling the entire object. Scrim results menu is now also displayed this way. Adjusted flowcharts for Day 1 and 2 to account for these changes. Next: see if there is a better way to get the single instance of the scrim results menu in the scene.